+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+ - Added `~=` operator for `Regex`
+
 ## 2.3.0
 
  - Added `appendInterpolation(_:format:)` and `appendInterpolation(_:if:)` string interpolation methods